feat: queue cmd for AI
